day-trader/requirements.txt
alpaca-py==0.28.1
fastapi==0.110.3
uvicorn==0.30.6
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
pydantic==2.8.2
httpx==0.27.2
